Refactor events/favorites list
The EventsView and the FavoritesView use the same design.
Make one view to share between the two views.

Also, reuse the eventModel for the favorite view, filtered
by the favorite property being true. This make the updates of the models
easier to handle.

Add a delegate for the view for better legibility of the code.

Change-Id: Id3c231b7e720311701b73f216895c774f4c93009
Reviewed-by: Niels Weber <niels.weber@digia.com>