Update only the eventModel when updating the favorites
The favoriteModel is only used at startup to initialize the
favorite property of each event element so we don't need to keep
it updated.

The favorite objects from the server side are however updated as
expected with the create and remove queries.

With this change we then don't need to read all the favorites values
from the server each time a favorite is updated.

Change-Id: I10e042403ff31a255ee8874b6ce3264b4e2f0f9e
Reviewed-by: Niels Weber <niels.weber@digia.com>