Disable dynamicSortFilter for the proxy model
There are some models issues when adding/removing favorite events from
the Presentations page.

Adding/removing favorites updates the eventModel. The eventModel
is sometimes used as source model for the SortProxyModel, as in the
EventsList. As stated by the documentation, dynamicSortFilter should be
set to false in that case so that the proxy model is not automatically
sorted every time the model change. In our case, only the favorite field
is updated so there is no need to call sort() again.

Change-Id: I01a752748a88f4033349b9cb7af5dba47e8f34ae
Reviewed-by: Niels Weber <niels.weber@digia.com>