Enable some caching of events/tracks/breaks/days
For the current conference. Load the data from the file first and
then check for updated data from the backend. If the data loaded
is different from the data queried, update the model, do nothing otherwise.

Don't save the favorites, the favorites are only updated on the server side.
Don't save the time list model since it can change depending of the day
selected.

Change-Id: Id55a6fb613d7e411c5e289b6dc2edb66ae7377b4
Reviewed-by: Niels Weber <niels.weber@digia.com>